Added files via upload
This is a stock analyzer with a GUI written in Python. This program will take any ticker you enter into and evaluate it against Warren Buffet's rules of investing. The information is grabbed from morningstar and yahoo finance. After analyzing its finds it will then make a recommendation as to whether you should purchase that stock.